[FIX] website_sale_stock : Remove import website_sale from variant_mixin file
Current behavior before PR:
While working on this https://github.com/odoo/enterprise/pull/74813/commits/337ea18baebdf4c896bc3e503448207fd9095991 we were overriding
a method in VariantMixin but it was shadowed and not executed this was
happening because of the order JS is loading the files when importing
website_sale.

Desired behavior after PR is merged:
We are removing the import of website_sale from variant_mixin.js
and moving the overridden method to website_sale.js to avoid having
this problem in the future.

X-original-commit: 3e553bb1993538ed0d2628e953db55054d26eea0
