+++ addons/website_sale_stock/__manifest__.py
@@ -30,9 +30,9 @@ Then it can be made specific at the product level.
     'auto_install': True,
     'assets': {
         'web.assets_frontend': [
+            ('before', 'website_sale/static/src/js/website_sale.js', 'website_sale_stock/static/src/js/variant_mixin.js'),
             'website_sale_stock/static/src/js/product/*',
             'website_sale_stock/static/src/js/product_configurator_dialog/*',
-            'website_sale_stock/static/src/js/variant_mixin.js',
             'website_sale_stock/static/src/js/website_sale.js',
             'website_sale_stock/static/src/js/website_sale_reorder.js',
             'website_sale_stock/static/src/xml/**/*',
